third_party/WebKit/LayoutTests/media/media-controller-effective-playback-rate-expected.txt

EVENT(canplay)
EXPECTED (video.controller.playbackRate == '1') OK
EXPECTED (video.playbackRate == '0') OK
EVENT(seeked)
RUN(video.play())
EVENT(play)
EVENT(timeupdate)
EXPECTED (video.currentTime != '0') OK
END OF TEST
